rewrite to set float values to SOC

+++ src/daly-bms-uart.h
@@ -261,7 +261,7 @@ public:
      * @brief set the SOC
      *
      */
-    bool setSOC(uint16_t sw);
+    bool setSOC(float sw);
 
     /**
      * @brief Read the charge and discharge MOS States
